services/nlp-service/requirements.txt
kafka-python
spacy
pydantic
scikit-learn
numpy
